requirements.txt
gunicorn
plotly
dash>=2.0
dash_bootstrap_components
dash_bio
pandas
numpy
